requirements.txt
opencv-python==4.8.0.76
mediapipe==0.10.8
numpy==1.24.4
pyautogui==0.9.54
face-recognition==1.3.0
keyboard==0.13.5
Pillow==10.2.0 